fixup: reorganize into protocol_engine.clients based on review feedback
