+++ api/src/opentrons/protocol_engine/clients/__init__.py
@@ -1,0 +1,5 @@
+"""ProtocolEngine clients."""
+from .sync_client import SyncClient
+from .transports import AbstractSyncTransport, ChildThreadTransport
+
+__all__ = ["SyncClient", "AbstractSyncTransport", "ChildThreadTransport"]
