+++ api/src/opentrons/protocols/implementations/engine/__init__.py
@@ -2,6 +2,5 @@
 
 from .protocol_context import ProtocolEngineContext
 from .labware_context import LabwareContext
-from .sync_engine import SyncProtocolEngine
 
-__all__ = ["ProtocolEngineContext", "LabwareContext", "SyncProtocolEngine"]
+__all__ = ["ProtocolEngineContext", "LabwareContext"]
