+++ api/src/opentrons/protocols/implementations/engine/protocol_context.py
@@ -5,7 +5,8 @@ from typing import Optional, Dict
 
 from opentrons import types
 from opentrons.hardware_control import API as HardwareAPI
-from opentrons.protocol_engine import ProtocolEngine, StateView, DeckSlotLocation
+from opentrons.protocol_engine import StateView, DeckSlotLocation
+from opentrons.protocol_engine.clients import SyncClient as ProtocolEngineClient
 from opentrons.protocols.api_support.util import HardwareManager, AxisMaxSpeeds
 from opentrons.protocols.geometry.deck import Deck
 from opentrons.protocols.geometry.deck_item import DeckItem
@@ -17,37 +18,21 @@ from opentrons.protocols.implementations.interfaces.protocol_context import \
     ProtocolContextInterface, InstrumentDict, LoadModuleResult
 from opentrons_shared_data.labware.dev_types import LabwareDefinition
 
-from .sync_engine import SyncProtocolEngine
 from .labware_context import LabwareContext
 
 
 class ProtocolEngineContext(ProtocolContextInterface):
     """ProtocolContextInterface that interacts with Protocol Engine"""
 
-    _sync_engine: SyncProtocolEngine
+    _client: ProtocolEngineClient
     _state_view: StateView
 
     def __init__(
         self,
-        sync_engine: SyncProtocolEngine,
-        state_view: StateView,
+        client: ProtocolEngineClient
     ) -> None:
         """Initialize a ProtocolEngineContext. Prefer `create` classmethod."""
-        self._sync_engine = sync_engine
-        self._state_view = state_view
-
-    @classmethod
-    def create(cls, protocol_engine: ProtocolEngine) -> ProtocolEngineContext:
-        """Create a ProtocolEngine context.
-
-        Args:
-            protocol_engine: A ProtocolEngine to manage protocol state and
-                handle command execution.
-
-        Returns:
-            A ready to use, concrete ProtocolContextInterface.
-        """
-        raise NotImplementedError()
+        self._client = client
 
     def get_bundled_data(self) -> Dict[str, bytes]:
         raise NotImplementedError()
@@ -91,14 +76,17 @@ class ProtocolEngineContext(ProtocolContextInterface):
         version: Optional[int] = None,
     ) -> LabwareInterface:
         # TODO(mc, 2021-03-04): handle optional namespace and version
-        result = self._sync_engine.load_labware(
+        result = self._client.load_labware(
             load_name=load_name,
             location=DeckSlotLocation(slot=types.DeckSlotName.from_primitive(location)),
             namespace=namespace,  # type: ignore[arg-type]
             version=version,  # type: ignore[arg-type]
         )
 
-        return LabwareContext(labware_id=result.labwareId, state_view=self._state_view)
+        return LabwareContext(
+            labware_id=result.labwareId,
+            state_view=self._client.state,
+        )
 
     def load_module(self, module_name: str,
                     location: Optional[types.DeckLocation] = None,
